+++ frontend/TaskActions/ActionRefuse.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace TaskForce;
+namespace frontend\TaskActions;
 
-class ActionRespond extends AbstractAction
+class ActionRefuse extends AbstractAction
 {
-    public const ACTION_CODE = 10;
-    private const ACTION_NAME = "Откликнутся";
+    public const ACTION_CODE = 30;
+    private const ACTION_NAME = "Отказаться";
+
     public function __construct()
     {
         $this->actionName = self::ACTION_NAME;
