+++ frontend/TaskActions/ActionRespond.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace TaskForce;
+namespace frontend\TaskActions;
 
-class ActionRefuse extends AbstractAction
+class ActionRespond extends AbstractAction
 {
-    public const ACTION_CODE = 30;
-    private const ACTION_NAME = "Отказаться";
-
+    public const ACTION_CODE = 10;
+    private const ACTION_NAME = "Откликнутся";
     public function __construct()
     {
         $this->actionName = self::ACTION_NAME;
