+++ frontend/views/components/actionButtons.php
@@ -1,0 +1,11 @@
+<div class="content-view__action-buttons">
+    <button class=" button button__big-color response-button open-modal"
+            type="button" data-for="response-form">Откликнуться
+    </button>
+    <button class="button button__big-color refusal-button open-modal"
+            type="button" data-for="refuse-form">Отказаться
+    </button>
+    <button class="button button__big-color request-button open-modal"
+            type="button" data-for="complete-form">Завершить
+    </button>
+</div>
