+++ frontend/views/components/chat.php
@@ -1,0 +1,25 @@
+<div class="connect-desk__chat">
+    <h3>Переписка</h3>
+    <div class="chat__overflow">
+        <div class="chat__message chat__message--out">
+            <p class="chat__message-time">10.05.2019, 14:56</p>
+            <p class="chat__message-text">Привет. Во сколько сможешь
+                приступить к работе?</p>
+        </div>
+        <div class="chat__message chat__message--in">
+            <p class="chat__message-time">10.05.2019, 14:57</p>
+            <p class="chat__message-text">На задание
+                выделены всего сутки, так что через час</p>
+        </div>
+        <div class="chat__message chat__message--out">
+            <p class="chat__message-time">10.05.2019, 14:57</p>
+            <p class="chat__message-text">Хорошо. Думаю, мы справимся</p>
+        </div>
+    </div>
+    <p class="chat__your-message">Ваше сообщение</p>
+    <form class="chat__form">
+            <textarea class="input textarea textarea-chat" rows="2" name="message-text"
+                      placeholder="Текст сообщения"></textarea>
+        <button class="button chat__button" type="submit">Отправить</button>
+    </form>
+</div>
